Replace "коммит" with "фиксация".
Sorry, guys, I was forced to make this change in order the branch to be merged.
I do not really think that using word "фиксация" makes the translation more
clear and understandable. Also I do not think it makes the text "more russian".

Feel free to revert this change as soon as readership feedback arrive.